+++ app/src/main/assets/about.properties
@@ -1,6 +1,6 @@
 about.version=2.0.0-rc.5
-about.build=2986
-about.date=Thu Nov 29 10:54:09 2018
+about.build=3022
+about.date=Thu Nov 29 17:40:32 2018
 about.commit=
 about.commitFull=
 about.github=https://github.com/flyve-mdm/flyve-mdm-android-agent
